third_party/blink/web_tests/external/wpt/html/browsers/browsing-the-web/history-traversal/srcdoc/consecutive-srcdoc-expected.txt

This is a testharness.js-based test.
[FAIL] changing srcdoc to about:srcdoc#yo then another srcdoc does two push navigations and we can navigate back
  assert_equals: srcdoc content must be restored from history expected "srcdoc1" but got "srcdoc2"
Harness: the test ran to completion.
